Created PassengerController Class and move code from PassengerJpaApplication into it
