+++ src/main/java/ie/atu/PassengerJPA/PassengerJpaApplication.java
@@ -2,35 +2,11 @@ package ie.atu.PassengerJPA;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
 
-import java.util.List;
 
-@RestController
 @SpringBootApplication
-@RequestMapping(path="api/passenger")
 public class PassengerJpaApplication {
-
-	PassengerService myService;
-
-	public PassengerJpaApplication(PassengerService myService) {
-		this.myService = myService;
-	}
-
 	public static void main(String[] args) {
 		SpringApplication.run(PassengerJpaApplication.class, args);
-	}
-
-	@GetMapping
-	public List<Passenger> getPassengers(){
-		return myService.getPassengers();
-	}
-
-	@GetMapping("/{passengerID}")
-	public Passenger getPassenger(@PathVariable String passengerID){
-		return myService.getPassenger(passengerID);
 	}
 }
